+++ internal/effort/safeio_darwin.go
@@ -1,0 +1,15 @@
+//go:build darwin
+
+package effort
+
+import (
+	"os"
+	"syscall"
+)
+
+func linkCount(info os.FileInfo) uint64 {
+	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
+		return uint64(stat.Nlink)
+	}
+	return 1 // Darwin os.FileInfo values carry syscall.Stat_t.
+}